refactor(validation): move requests into types folder
