+++ src/types/transport/requests/files/GetFilesRequest.ts
@@ -1,4 +1,4 @@
-import { OwnerRestriction } from "../../../common";
+import { OwnerRestriction } from "../../../useCases/common";
 
 export interface GetFilesRequest {
     query?: any;
